+++ premium/web-frontend/modules/baserow_premium/plugin.js
@@ -1,7 +1,8 @@
 import { PremPlugin } from '@baserow_premium/plugins'
-import { UsersAdminType } from '@baserow_premium/adminTypes'
+import { DashboardType, UsersAdminType } from '@baserow_premium/adminTypes'
 
 export default ({ app }) => {
   app.$registry.register('plugin', new PremPlugin())
+  app.$registry.register('admin', new DashboardType())
   app.$registry.register('admin', new UsersAdminType())
 }