+++ premium/web-frontend/modules/baserow_premium/routes.js
@@ -1,8 +1,11 @@
-// import path from 'path'
-
 import path from 'path'
 
 export const routes = [
+  {
+    name: 'admin-dashboard',
+    path: '/admin/dashboard',
+    component: path.resolve(__dirname, 'pages/admin/dashboard.vue'),
+  },
   {
     name: 'admin-users',
     path: '/admin/users',