+++ changelog.md
@@ -9,6 +9,7 @@
 * Made it possible to order the groups by drag and drop.
 * Made it possible to order the applications by drag and drop.
 * Made it possible to order the tables by drag and drop.
+* **Premium**: Added an admin dashboard.
 * Added today, this month and this year filter.
 * Added a human-readable error message when a user tries to sign in with a deactivated
   account.